[Console] Clean up tests
