+++ src/plugins/console/public/application/containers/editor/monaco/hooks/use_resize_checker_utils.ts
