+++ src/plugins/console/public/application/containers/editor/monaco/hooks/use_setup_autocomplete_polling.ts
@@ -7,7 +7,7 @@
  */
 
 import { useEffect } from 'react';
-import { AutocompleteInfo, Settings } from '../../../../services';
+import { AutocompleteInfo, Settings } from '../../../../../services';
 
 interface SetupAutocompletePollingParams {
   /** The Console autocomplete service. */
