+++ src/plugins/console/public/application/containers/editor/monaco/utils/autocomplete_utils.test.ts
@@ -1,0 +1,68 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+/*
+ * Mock the function "populateContext" that accesses the autocomplete definitions
+ */
+const mockPopulateContext = jest.fn();
+
+jest.mock('../../../../../lib/autocomplete/engine', () => {
+  return {
+    populateContext: (...args: any) => {
+      mockPopulateContext(args);
+    },
+  };
+});
+import { AutoCompleteContext } from '../../../../../lib/autocomplete/types';
+import { getDocumentationLinkFromAutocomplete } from './autocomplete_utils';
+
+describe('autocomplete_utils', () => {
+  describe('getDocumentationLinkFromAutocomplete', () => {
+    const mockRequest = { method: 'GET', url: '_search', data: [] };
+    const version = '8.13';
+    const expectedLink = 'http://elastic.co/8.13/_search';
+
+    it('correctly replaces {branch} with the version', () => {
+      const endpoint = {
+        documentation: 'http://elastic.co/{branch}/_search',
+      } as AutoCompleteContext['endpoint'];
+      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
+      mockPopulateContext.mockImplementation((...args) => {
+        const context = args[0][1];
+        context.endpoint = endpoint;
+      });
+      const link = getDocumentationLinkFromAutocomplete(mockRequest, version);
+      expect(link).toBe(expectedLink);
+    });
+
+    it('correctly replaces /master/ with the version', () => {
+      const endpoint = {
+        documentation: 'http://elastic.co/master/_search',
+      } as AutoCompleteContext['endpoint'];
+      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
+      mockPopulateContext.mockImplementation((...args) => {
+        const context = args[0][1];
+        context.endpoint = endpoint;
+      });
+      const link = getDocumentationLinkFromAutocomplete(mockRequest, version);
+      expect(link).toBe(expectedLink);
+    });
+
+    it('correctly replaces /current/ with the version', () => {
+      const endpoint = {
+        documentation: 'http://elastic.co/current/_search',
+      } as AutoCompleteContext['endpoint'];
+      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
+      mockPopulateContext.mockImplementation((...args) => {
+        const context = args[0][1];
+        context.endpoint = endpoint;
+      });
+      const link = getDocumentationLinkFromAutocomplete(mockRequest, version);
+      expect(link).toBe(expectedLink);
+    });
+  });
+});
