+++ src/plugins/console/public/application/containers/editor/monaco/utils/index.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { AutocompleteType, SELECTED_REQUESTS_CLASSNAME } from './constants';
+export {
+  getRequestStartLineNumber,
+  getRequestEndLineNumber,
+  stringifyRequest,
+  replaceRequestVariables,
+  getCurlRequest,
+  trackSentRequests,
+} from './requests_utils';
+export {
+  getDocumentationLinkFromAutocomplete,
+  getMethodCompletionItems,
+  getUrlPathCompletionItems,
+  getUrlParamsCompletionItems,
+  getBodyCompletionItems,
+} from './autocomplete_utils';
