+++ src/plugins/console/public/application/containers/editor/monaco/utils/requests_utils.test.ts
@@ -7,31 +7,14 @@
  */
 
 import {
-  getBodyTokenPath,
   getCurlRequest,
-  getDocumentationLink,
-  removeTrailingWhitespaces,
   replaceRequestVariables,
   stringifyRequest,
   trackSentRequests,
-} from './utils';
-import { MetricsTracker } from '../../../../types';
-import { AutoCompleteContext } from '../../../../lib/autocomplete/types';
+} from './requests_utils';
+import { MetricsTracker } from '../../../../../types';
 
-/*
- * Mock the function "populateContext" that accesses the autocomplete definitions
- */
-const mockPopulateContext = jest.fn();
-
-jest.mock('../../../../lib/autocomplete/engine', () => {
-  return {
-    populateContext: (...args: any) => {
-      mockPopulateContext(args);
-    },
-  };
-});
-
-describe('monaco editor utils', () => {
+describe('requests_utils', () => {
   const dataObjects = [
     {
       query: {
@@ -42,24 +25,6 @@ describe('monaco editor utils', () => {
       test: 'test',
     },
   ];
-  describe('removeTrailingWhitespaces', () => {
-    it(`works with an empty string`, () => {
-      const url = '';
-      const result = removeTrailingWhitespaces(url);
-      expect(result).toBe(url);
-    });
-    it(`doesn't change the string if no trailing whitespaces`, () => {
-      const url = '_search';
-      const result = removeTrailingWhitespaces(url);
-      expect(result).toBe(url);
-    });
-    it(`removes any text after the first whitespace`, () => {
-      const url = '_search some_text';
-      const result = removeTrailingWhitespaces(url);
-      expect(result).toBe('_search');
-    });
-  });
-
   describe('stringifyRequest', () => {
     const request = {
       startOffset: 0,
@@ -194,139 +159,5 @@ describe('monaco editor utils', () => {
       expect(mockMetricsTracker.count).toHaveBeenNthCalledWith(1, 'GET__search');
       expect(mockMetricsTracker.count).toHaveBeenNthCalledWith(2, 'POST__test');
     });
-  });
-
-  describe('getDocumentationLink', () => {
-    const mockRequest = { method: 'GET', url: '_search', data: [] };
-    const version = '8.13';
-    const expectedLink = 'http://elastic.co/8.13/_search';
-
-    it('correctly replaces {branch} with the version', () => {
-      const endpoint = {
-        documentation: 'http://elastic.co/{branch}/_search',
-      } as AutoCompleteContext['endpoint'];
-      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
-      mockPopulateContext.mockImplementation((...args) => {
-        const context = args[0][1];
-        context.endpoint = endpoint;
-      });
-      const link = getDocumentationLink(mockRequest, version);
-      expect(link).toBe(expectedLink);
-    });
-
-    it('correctly replaces /master/ with the version', () => {
-      const endpoint = {
-        documentation: 'http://elastic.co/master/_search',
-      } as AutoCompleteContext['endpoint'];
-      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
-      mockPopulateContext.mockImplementation((...args) => {
-        const context = args[0][1];
-        context.endpoint = endpoint;
-      });
-      const link = getDocumentationLink(mockRequest, version);
-      expect(link).toBe(expectedLink);
-    });
-
-    it('correctly replaces /current/ with the version', () => {
-      const endpoint = {
-        documentation: 'http://elastic.co/current/_search',
-      } as AutoCompleteContext['endpoint'];
-      // mock the populateContext function that finds the correct autocomplete endpoint object and puts it into the context object
-      mockPopulateContext.mockImplementation((...args) => {
-        const context = args[0][1];
-        context.endpoint = endpoint;
-      });
-      const link = getDocumentationLink(mockRequest, version);
-      expect(link).toBe(expectedLink);
-    });
-  });
-
-  describe('getBodyTokenPath', () => {
-    const testCases: Array<{ value: string; tokens: string[] }> = [
-      {
-        // add opening curly brackets to tokens
-        value: '{',
-        tokens: ['{'],
-      },
-      {
-        // allow whitespaces
-        value: '  {',
-        tokens: ['{'],
-      },
-      {
-        // allow line comments
-        value: '//comment\n{',
-        tokens: ['{'],
-      },
-      {
-        // inside the object line comment are ignored
-        value: '{//comment',
-        tokens: ['{'],
-      },
-      {
-        value: '{//comment\n"test":',
-        tokens: ['{', 'test'],
-      },
-      {
-        // do not add property name if no colon (:)
-        value: '{"test"',
-        tokens: ['{'],
-      },
-      {
-        // add property names to tokens (double quotes are removed)
-        value: '{"test":',
-        tokens: ['{', 'test'],
-      },
-      {
-        // add nested object to tokens
-        value: '{"test":{',
-        tokens: ['{', 'test', '{'],
-      },
-      {
-        // empty object
-        value: '{}',
-        tokens: [],
-      },
-      {
-        // empty object with inline comment
-        value: '{//comment\n}',
-        tokens: [],
-      },
-      {
-        value: '{"test":[',
-        tokens: ['{', 'test', '['],
-      },
-      {
-        value: '{"test":123,',
-        tokens: ['{'],
-      },
-      {
-        value: '{"test":{},',
-        tokens: ['{'],
-      },
-      {
-        value: '{"test":[],',
-        tokens: ['{'],
-      },
-      {
-        value: '{"property1":"value1","property2":',
-        tokens: ['{', 'property2'],
-      },
-      {
-        value: '{"property1":[123,',
-        tokens: ['{', 'property1', '['],
-      },
-      {
-        value: '{"property1":[123,"test"]',
-        tokens: ['{', 'property1'],
-      },
-    ];
-    for (const testCase of testCases) {
-      const { value, tokens } = testCase;
-      it(`${value}`, () => {
-        const parsedTokens = getBodyTokenPath(value);
-        expect(parsedTokens).toEqual(tokens);
-      });
-    }
   });
 });
