+++ src/plugins/console/public/application/hooks/use_send_current_request/index.ts
@@ -7,3 +7,4 @@
  */
 
 export { useSendCurrentRequest } from './use_send_current_request';
+export { sendRequest } from './send_request';
